implement creation of CDI concurrency beans based on configuration
The configuration is done during previous scanning of annotations
and the result is shared via JNDI. During BeforeBeanDiscovery event
this information is retrieved and required beans are registered.

Unfortunately, it's not possible to use
@Observes @WithAnnotations(ContextServiceDefinition.class),
because this annotation doesn't need to be on a CDI bean, e.g. it is
not recognized.